backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
pydantic-settings==2.6.1
email-validator>=2.0.0
